:recycle: sessionStudent 삭제, 로직 SessionEnrollment으로 이동
