--- src/main/java/nextstep/courses/domain/session/SessionStudent.java
@@ -1,25 +1,0 @@
-package nextstep.courses.domain.session;
-
-import nextstep.users.domain.NsUser;
-
-public class SessionStudent {
-
-    private int maxStudentCount;
-
-    private SessionStudents sessionStudents;
-
-    public SessionStudent(int maxStudentCount, SessionStudents sessionStudents) {
-        this.maxStudentCount = maxStudentCount;
-        this.sessionStudents = sessionStudents;
-    }
-
-    public void isUnderMaxStudentCount() {
-        if (sessionStudents.enrolledStudentsCount() >= maxStudentCount) {
-            throw new IllegalArgumentException("수강 인원이 다 찼습니다");
-        }
-    }
-
-    public void add(NsUser student) {
-        sessionStudents.add(student);
-    }
-}
